Use ptrdiff_t for pointer differences and size_t for all offsets into buffers that are initialized with pointer differences.  Based on a patch from Jacek Konieczny and necessary for AMD64 support.

+++ innd/chan.c
@@ -640,7 +640,7 @@ WCHANsetfrombuffer(CHANNEL *cp, struct buffer *bp)
 int
 CHANreadtext(CHANNEL *cp)
 {
-    int	                i, j;
+    ptrdiff_t           i, j;
     struct buffer       *bp;
     char		*p;
     int			oerrno;
@@ -918,7 +918,7 @@ CHANreadloop(void)
 {
     static char		EXITING[] = "INND exiting because of signal\n";
     static int		fd;
-    int			i, j;
+    ptrdiff_t		i, j;
     int			startpoint;
     int			count;
     int			lastfd;
